Fix #8 - crash when file is deleted
preform a invalidation step on LocalStorage before getting model to use
in player, that way we are sure that all file path in model are valid in
can be assumed as existeing file. altough it is expected to crash when
user delete a file after opening player and then tries to play it.